Services implemented
I have implemented the missing services and service impl classes. Onwards to creating tests for the remaining services.

+++ TCGTradingPost/src/main/java/com/tcgtp/repositories/OrderStatusRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.tcgtp.domain.OrderStatus;
 
-public interface OrderStatusRepository extends CrudRepository<OrderStatus, Integer>{
+public interface OrderStatusRepository extends CrudRepository<OrderStatus, Long>{
 
 }
